% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values_rebuilder.R
\name{get_values}
\alias{get_values}
\title{Pull output from Rebuilder program RES.CSV fle}
\usage{
get_values(dir, num_rows = 200)
}
\arguments{
\item{dir}{Directory where rebuilder output
files are stored.}

\item{num_rows}{Number of columns in the csv file to be read}
}
\description{
Pull output from Rebuilder program RES.CSV fle
}
\author{
Chantel Wetzel
}
